requirements.txt
fastapi~=0.115.5
uvicorn
streamlit~=1.40.1
pandas~=2.2.3
scikit-learn
statsmodels
plotly
matplotlib
seaborn
openpyxl
gradio
gradio-client
python-multipart
httpx~=0.27.2
qrcode~=8.0
pillow~=11.0.0
pyzbar
langchain
replicate
pywhatkit
faiss-cpu
langchain_community
torch
sentence_transformers
pushbullet.py
langchain-huggingface
langchain-openai
langgraph
requests